More about Endpoint

+++ MovieAPI/src/main/java/dev/Innocent/DTO/MoviePageResponse.java
@@ -10,6 +10,6 @@ public record MoviePageResponse(
         Integer currentPage,
         boolean isLast,
         boolean isFirst,
-        int totalElements
+        Long totalElements
 ) {
 }
